requirements.txt
opencv-python==4.6.0.66
python-dotenv==0.21.0
slack-sdk==3.19.1
